EDGEucatorWebApp/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai==1.14.0
httpx==0.27.2
python-dotenv==1.0.1
# gesture_server.py
mediapipe>=0.10.0
opencv-python>=4.9.0
websockets>=12.0
# wake word (Picovoice Porcupine) — must match .ppn version (v3)
pvporcupine>=3.0.0,<4.0.0
pvrecorder>=1.2.0
